Release candidate

+++ contracts/Nodes.sol
@@ -310,8 +310,6 @@ contract Nodes is Permissions {
 
         _setNodeLeft(nodeIndex);
         _deleteNode(nodeIndex);
-        
-        BountyV2(contractManager.getBounty()).handleNodeRemoving(nodes[nodeIndex].validatorId);
 
         emit ExitCompleted(
             nodeIndex,
@@ -343,7 +341,7 @@ contract Nodes is Permissions {
         validatorToNodeIndexes[validatorId].pop();
         address nodeOwner = _publicKeyToAddress(nodes[nodeIndex].publicKey);
         if (validatorService.getValidatorIdByNodeAddress(nodeOwner) == validatorId) {
-            if (nodeIndexes[nodeOwner].numberOfNodes == 1) {
+            if (nodeIndexes[nodeOwner].numberOfNodes == 1 && !validatorService.validatorAddressExists(nodeOwner)) {
                 validatorService.removeNodeAddress(validatorId, nodeOwner);
             }
             nodeIndexes[nodeOwner].isNodeExist[nodeIndex] = false;
@@ -445,18 +443,6 @@ contract Nodes is Permissions {
         }
         require(permitted, "Sender is not permitted to call this function");
         _setNodeActive(nodeIndex);
-    }
-
-    function populateBountyV2(uint from, uint to) external onlyOwner {
-        BountyV2 bounty = BountyV2(contractManager.getBounty());
-        uint nodeCreationWindow = bounty.nodeCreationWindowSeconds();
-        bounty.setNodeCreationWindowSeconds(uint(-1) / 2);
-        for (uint nodeId = from; nodeId < _min(nodes.length, to); ++nodeId) {
-            if (nodes[nodeId].status != NodeStatus.Left) {
-                bounty.handleNodeCreation(nodes[nodeId].validatorId);
-            }
-        }
-        bounty.setNodeCreationWindowSeconds(nodeCreationWindow);
     }
 
     /**
@@ -873,8 +859,6 @@ contract Nodes is Permissions {
         }));
         spaceToNodes[constantsHolder.TOTAL_SPACE_ON_NODE()].push(nodeIndex);
         numberOfActiveNodes++;
-        
-        BountyV2(contractManager.getBounty()).handleNodeCreation(validatorId);
     }
 
     /**
